RegulAgentV2/regulagent-backend/requirements/base.txt
# Core framework
Django>=4.2,<5.1
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.0
django-simple-history>=3.4
jsonpatch>=1.33  # For plan version diffs (RFC 6902)

# Database and cache
psycopg2-binary>=2.9
django-redis>=5.4
redis>=5.0
hiredis>=2.2
django-tenants>=3.7.0
django-tenant-users>=1.4.0

# Celery
celery>=5.3,<5.4
django-celery-beat>=2.5

# Storage & files
django-storages>=1.14
boto3>=1.34
pillow>=10.0

# Config and server
python-dotenv>=1.0
gunicorn>=21.2
uvicorn>=0.30
PyYAML>=6.0

# Vector similarity (pgvector with Django integration module)
pgvector>=0.2.5

# Parsing & OCR stack
pdfplumber>=0.11.7
pdf2image>=1.17
PyMuPDF>=1.24  # fitz
camelot-py[base]>=0.11  # tables from PDFs (requires ghostscript)
tabula-py>=2.9  # tables via Java Tabula (requires Java runtime)
pytesseract>=0.3.10  # requires system Tesseract
opencv-python-headless>=4.10  # used by camelot and image pre-processing
pandas>=2.2
numpy>=1.26

# Cloud OCR providers
# AWS Textract via boto3 (already included above)
google-cloud-documentai>=2.28
google-auth>=2.29

# Web automation (RRC connector)
playwright>=1.46

# HTML parsing for policy ingestion
beautifulsoup4>=4.12
requests>=2.32

openai>=1.6.0,<2.0.0


